+++ app/src/main/java/org/simple/clinic/summary/prescribeddrugs/DrugSummaryUiRenderer.kt
@@ -1,0 +1,14 @@
+package org.simple.clinic.summary.prescribeddrugs
+
+import org.simple.clinic.mobius.ViewRenderer
+
+class DrugSummaryUiRenderer(
+    private val ui: DrugSummaryUi
+) : ViewRenderer<DrugSummaryModel> {
+
+  override fun render(model: DrugSummaryModel) {
+    if (model.hasPrescribedDrugs) {
+      ui.populatePrescribedDrugs(model.prescribedDrugs!!)
+    }
+  }
+}